requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
stripe==7.7.0
python-dotenv==1.0.0
mautrix==0.20.3
aiohttp==3.9.1
gspread==5.12.4
google-auth==2.25.2
